Character Controller Added
I downloaded a free First Person Character controller from the Unity Asset Store, I tried importing this without the unnecessary files such as the mobile folder, but this caused many errors with my project and prevented it from launching outside of Safe Mode. I had to uninstall and re-import the package with everything except for the environment to get this to work. I also had to remove the URP package that it came with to stop getting errors. Fortunately I can now walk around the map and explore as a player would.