Imported Textures and updated Terrain Data and Texture Data scripts
I imported some textures to be used on the terrain. I made it so that these textures have multiple properties such as set height, blend amount. tint etc, so they can be fully customisable in the editor. These textures are applied to the map mesh and we can see the map looking more polished. The next step is mainly bug fixes and refactoring.